EE-16064 Fix misnamed tests
